+++ examples/Ethernet/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,13 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  
+   mDNS library to support mDNS (registering services) and DNS-SD (service discovery).
+  
+  Based on and modified from https://github.com/arduino-libraries/ArduinoMDNS
+  Built by Khoi Hoang https://github.com/khoih-prog/MDNS_Generic
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"