+++ examples/WiFi/WiFiMultiFileProject/multiFileProject.h
@@ -1,0 +1,19 @@
+/****************************************************************************************************************************
+  multiFileProject.h
+  
+   mDNS library to support mDNS (registering services) and DNS-SD (service discovery).
+  
+  Based on and modified from https://github.com/arduino-libraries/ArduinoMDNS
+  Built by Khoi Hoang https://github.com/khoih-prog/MDNS_Generic
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#pragma once
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "MDNS_Generic.hpp"
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "MDNS_Responder.hpp"